refactor: Split business logic into hook
