+++ frontend/src/pages/login/LoginPage.tsx
@@ -1,8 +1,3 @@
-import { useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { SubmitHandler, useForm } from 'react-hook-form';
-import { yupResolver } from '@hookform/resolvers/yup';
-import { IUser } from '../../types';
 import {
   Background,
   ChildTemplate,
@@ -10,10 +5,8 @@ import {
   LoginForm,
   ParentTemplate,
   UserForm,
-  loginSchema,
 } from '../../components';
-import { useAppSelector } from '../../hooks/hooks';
-import { selectUser, useUserAuthenticatedMutation } from '../../features';
+import { useLoginPage } from './LoginPage.hook';
 
 export type LoginData = {
   userEmail: string;
@@ -21,36 +14,8 @@ export type LoginData = {
 };
 
 export default function LoginPage() {
-  const navigate = useNavigate();
-  const userInfo: IUser = useAppSelector(selectUser);
-
-  // Login via api
-  const [userAuthenticate, { error: loginError }] =
-    useUserAuthenticatedMutation();
-
-  // Submit login
-  const loginSubmit: SubmitHandler<LoginData> = (data: LoginData) => {
-    userAuthenticate({
-      email: data.userEmail,
-      password: data.userPassword,
-    });
-  };
-
-  const {
-    register,
-    handleSubmit,
-    formState: { errors: inputErrors },
-  } = useForm<LoginData>({
-    mode: 'onChange',
-    resolver: yupResolver(loginSchema),
-  });
-
-  // Go to home page after login
-  useEffect(() => {
-    if (userInfo.token && userInfo._id) {
-      navigate('/');
-    }
-  }, [userInfo]);
+  const [inputErrors, handleSubmit, loginSubmit, register, loginError] =
+    useLoginPage();
 
   return (
     <Layout>
